Make all operations exact up to 1 ULP
The use of a 'maximum relative error' for each type in the tests was a
bit of a hack and overshadowed the true necessity of coming within 1
ULP of a mostly exact reference result casted to the respective takum
type directly.

This has been reached now for both logarithmic and linear takums by
fixing many small things here and there. The automatic tests really made
this possible; despite the work of setting up the test environment they
really are a gift that keeps on giving.

+++ src/csc.c
@@ -11,7 +11,7 @@
  */
 #define RESULT_FIXER_MACRO(arg, res) (res)
 
-static double
+static long double
 cscl(long double f)
 {
 	return 1.0l / sinl(f);
@@ -26,7 +26,7 @@ csc(double f)
 UTIL_UNARY_FLOAT_TAKUM_WRAPPER(csc, csc, RESULT_FIXER_MACRO)
 UTIL_UNARY_FLOAT_TAKUM_LINEAR_WRAPPER(csc, csc, RESULT_FIXER_MACRO)
 
-static double
+static long double
 csc_pi_timesl(long double f)
 {
 	return cscl(PI * f);
